==> MTECH-PROJ-master/2_tnsrflw_regression_custom_functions.ipynb <==
import tensorflow as tf
import numpy as np
import matplotlib.pyplot as plt

def data_generator(num_data):
    a = np.random.rand(num_data).astype('float32')
    b = np.random.rand(num_data).astype('float32')
    c = np.random.rand(num_data).astype('float32')
    d = np.random.rand(num_data).astype('float32')
    return a,b,c,d

def function1(a,b,c,d):
    return np.reshape(2*a+3*b+5*c+7*d, (-1, 1))

def function2(a,b,c,d):
    return np.sin(a)*np.cos(b)*np.sin(c)*np.cos(d)

def function3(a,b,c,d):
    return a + np.power(b,c)*d

def function4(a,b,c,d):
    return a*b*c*d
    
def train_test_split(input_matrix, Y):
    num_data = Y.shape[0]
    num_train = round(num_data*.9)
    num_test = round(num_data*.1)
    return input_matrix[:num_train], Y[:num_train], input_matrix[-num_test:], Y[-num_test:]
    
def tnsrflw(train_input, train_output, test_input, test_output):
    # Function 1
#     k = tf.Variable(tf.random_normal([1, 4]))
#     A = tf.placeholder("float", [4, None])
#     Y = tf.placeholder("float")
#     pred = tf.matmul(k, A)
    
    # General Function 1
    c1 = tf.Variable(tf.random_normal([1]))
    c2 = tf.Variable(tf.random_normal([1]))
    c3 = tf.Variable(tf.random_normal([1]))
    c4 = tf.Variable(tf.random_normal([1]))
    c5 = tf.Variable(tf.random_normal([1]))
    c6 = tf.Variable(tf.random_normal([1]))
    c7 = tf.Variable(tf.random_normal([1]))
    c8 = tf.Variable(tf.random_normal([1]))
    c9 = tf.Variable(tf.random_normal([1]))
    c10 = tf.Variable(tf.random_normal([1]))
    c11 = tf.Variable(tf.random_normal([1]))
    c12 = tf.Variable(tf.random_normal([1]))
    c13 = tf.Variable(tf.random_normal([1]))
    c14 = tf.Variable(tf.random_normal([1]))
    c15 = tf.Variable(tf.random_normal([1]))
    c16 = tf.Variable(tf.random_normal([1]))
    
    k1 = tf.Variable(tf.random_normal([1]))
    k2 = tf.Variable(tf.random_normal([1]))
    k3 = tf.Variable(tf.random_normal([1]))
    k4 = tf.Variable(tf.random_normal([1]))
    
    l1 = tf.Variable(tf.random_normal([1]))
    l2 = tf.Variable(tf.random_normal([1]))
    l3 = tf.Variable(tf.random_normal([1]))
    l4 = tf.Variable(tf.random_normal([1]))
    l5 = tf.Variable(tf.random_normal([1]))
    l6 = tf.Variable(tf.random_normal([1]))
    l7 = tf.Variable(tf.random_normal([1]))
    l8 = tf.Variable(tf.random_normal([1]))
    l9 = tf.Variable(tf.random_normal([1]))
    l10 = tf.Variable(tf.random_normal([1]))
    l11 = tf.Variable(tf.random_normal([1]))
    l12 = tf.Variable(tf.random_normal([1]))
    
    m1 = tf.Variable(tf.random_normal([1]))
    m2 = tf.Variable(tf.random_normal([1]))
    m3 = tf.Variable(tf.random_normal([1]))
    m4 = tf.Variable(tf.random_normal([1]))
    m5 = tf.Variable(tf.random_normal([1]))
    m6 = tf.Variable(tf.random_normal([1]))
    m7 = tf.Variable(tf.random_normal([1]))
    m8 = tf.Variable(tf.random_normal([1]))
    m9 = tf.Variable(tf.random_normal([1]))
    m10 = tf.Variable(tf.random_normal([1]))
    m11 = tf.Variable(tf.random_normal([1]))
    m12 = tf.Variable(tf.random_normal([1]))
    
    n1 = tf.Variable(tf.random_normal([1]))
    n2 = tf.Variable(tf.random_normal([1]))
    n3 = tf.Variable(tf.random_normal([1]))
    n4 = tf.Variable(tf.random_normal([1]))
    
    A = tf.placeholder("float")
    B = tf.placeholder("float")
    C = tf.placeholder("float")
    D = tf.placeholder("float")

#     Function 1
#     pred = c1 + c2 * tf.pow(A,k1) * tf.pow(B,k2) * tf.pow(C,k3) * tf.pow(D,k4) + c3 * tf.pow(A,k5) * tf.pow(B,k6) * tf.pow(C,k7) * tf.pow(D,k8) + c4 * tf.pow(A,k9) * tf.pow(B,k10) * tf.pow(C,k11) * tf.pow(D,k12) + c5 * tf.pow(A,k13) * tf.pow(B,k14) * tf.pow(C,k15) * tf.pow(D,k16)
    
#     Function 2
    pred = c1 + c2 * tf.pow(A,k1) + c3 * tf.pow(B,k2)
    pred += c4 * tf.pow(C,k3) + c5 * tf.pow(D,k4)
    
    pred += c6 * tf.pow(A,l1) * tf.pow(B,l2) + c7 * tf.pow(A,l3) * tf.pow(C,l4)
    pred += c8 * tf.pow(A,l5) * tf.pow(D,l6) + c9 * tf.pow(B,l7) * tf.pow(C,l8)
    pred += c10 * tf.pow(C,l9) * tf.pow(D,l10) + c11 * tf.pow(B,l11) * tf.pow(D,l12)
    
    pred += c12 * tf.pow(A,m1) * tf.pow(B,m2) * tf.pow(C,m3) + c13 * tf.pow(B,m4) * tf.pow(C,m5) * tf.pow(D,m6)
    pred += c14 * tf.pow(A,m7) * tf.pow(C,m8) * tf.pow(D,m9) + c15 * tf.pow(A,m10) * tf.pow(B,m11) * tf.pow(D,m12)
    
    pred += c16 * tf.pow(A,n1) * tf.pow(B,n2) * tf.pow(C,n3) * tf.pow(D,n4)
    
    Y = tf.placeholder("float")
    
    cost = tf.reduce_mean(tf.pow(pred - Y, 2))
    optimizer = tf.train.GradientDescentOptimizer(0.001).minimize(cost)
    
    init = tf.global_variables_initializer()
    
    sess = tf.Session()
    sess.run(init)
    step = 0
    for epoch in range(50):
        for (x, y) in zip(train_input, train_output):
            sess.run(optimizer, feed_dict={A:x[0],B:x[1],C:x[2],D:x[3], Y:y})

            if(step % 5000 == 0):
                print(sess.run(cost, feed_dict={A:train_input[:,0],B:train_input[:,1],C:train_input[:,2],D:train_input[:,3], Y:train_output}))
            step+=1
            
    y_pred = sess.run(pred, feed_dict={A:test_input[:,0],B:test_input[:,1],C:test_input[:,2],D:test_input[:,3]})
    sess.close()
    return y_pred


def plot_y_vs_y_predicted(y_pred, y_test):
    plt.scatter(y_test, y_pred)
    plt.show()

def main_func():
    num_data = 10000
    a,b,c,d = data_generator(num_data)
    input_matrix = np.stack((a, b, c, d), axis=-1)
    Y = function4(a,b,c,d)

    train_input, train_output, test_input, test_output = train_test_split(input_matrix, Y)
#     print (train_input.shape)
    y_pred = tnsrflw(train_input, train_output, test_input, test_output)
#     print (y_pred)
    plot_y_vs_y_predicted(y_pred, test_output)

# The Main Function
main_func()
